refactor: migrate with previous mission code
