+++ room-escape-app/room-escape-web-app/build.gradle
@@ -11,4 +11,11 @@ dependencies {
 
     implementation "org.springframework.boot:spring-boot-starter-web"
     implementation "org.springframework:spring-jdbc"
+
+    // For Auth
+    implementation 'io.jsonwebtoken:jjwt:0.9.1'
+    implementation 'net.rakugakibox.spring.boot:logback-access-spring-boot-starter:2.7.1'
+
+    testImplementation 'org.springframework.boot:spring-boot-starter-test'
+    testImplementation 'io.rest-assured:rest-assured:3.3.0'
 }
