+++ room-escape-app/room-escape-web-app/src/main/java/nextstep/app/web/auth/AuthenticationException.java
@@ -1,0 +1,4 @@
+package nextstep.app.web.auth;
+
+public class AuthenticationException extends RuntimeException {
+}
