+++ room-escape-app/room-escape-web-app/src/main/java/nextstep/app/web/auth/TokenRequest.java
@@ -1,0 +1,19 @@
+package nextstep.app.web.auth;
+
+public class TokenRequest {
+    private String username;
+    private String password;
+
+    public TokenRequest(String username, String password) {
+        this.username = username;
+        this.password = password;
+    }
+
+    public String getUsername() {
+        return username;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+}
