+++ room-escape-app/room-escape-web-app/src/main/java/nextstep/app/web/auth/TokenResponse.java
@@ -1,0 +1,16 @@
+package nextstep.app.web.auth;
+
+public class TokenResponse {
+    public String accessToken;
+
+    public TokenResponse() {
+    }
+
+    public TokenResponse(String accessToken) {
+        this.accessToken = accessToken;
+    }
+
+    public String getAccessToken() {
+        return accessToken;
+    }
+}
