+++ room-escape-app/room-escape-web-app/src/main/java/nextstep/app/web/reservation/ReservationWebResponse.java
@@ -28,6 +28,10 @@ class ReservationWebResponse {
         return id;
     }
 
+    public Long getScheduleId() {
+        return scheduleId;
+    }
+
     public LocalDate getDate() {
         return date;
     }
