+++ room-escape-core/build.gradle
@@ -4,7 +4,7 @@ plugins {
 
 dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-jdbc'
-    implementation 'com.h2database:h2'
+    runtimeOnly 'com.h2database:h2'
 }
 
 bootJar { enabled = false }
