+++ room-escape-core/src/main/java/nextstep/core/member/in/MemberUseCase.java
@@ -1,0 +1,7 @@
+package nextstep.core.member.in;
+
+public interface MemberUseCase {
+    MemberResponse register(MemberRegisterRequest request);
+
+    MemberResponse login(Long id);
+}
