+++ room-escape-core/src/main/resources/schema.sql
@@ -4,8 +4,8 @@ drop table if exists schedules;
 
 create table reservation
 (
-    id          bigint      not null primary key auto_increment comment '예약 아이디',
-    schedule_id bigint      not null comment '스케줄 아이디',
+    id          bigint      not null primary key auto_increment comment '예약 식별자',
+    schedule_id bigint      not null comment '스케줄 식별자',
     date        date        not null default now() comment '예약 날짜',
     time        time        not null default now() comment '예약 시간',
     name        varchar(32) not null comment '예약자 이름'
@@ -13,7 +13,7 @@ create table reservation
 
 create table themes
 (
-    id    bigint       not null primary key auto_increment comment '테마 아이디',
+    id    bigint       not null primary key auto_increment comment '테마 식별자',
     name  varchar(32)  not null comment '테마 이름',
     desc  varchar(128) not null comment '테마 설명',
     price bigint       not null comment '테마 가격'
@@ -21,8 +21,17 @@ create table themes
 
 create table schedules
 (
-    id       bigint not null primary key auto_increment comment '스케줄 아이디',
-    theme_id bigint not null comment '테마 아이디',
+    id       bigint not null primary key auto_increment comment '스케줄 식별자',
+    theme_id bigint not null comment '테마 식별자',
     date     date   not null default now() comment '스케줄 날짜',
     time     time   not null default now() comment '스케줄 시간'
 );
+
+create table member
+(
+    id       bigint      not null primary key auto_increment comment '회원 식별자',
+    username varchar(20) not null comment '회원 아이디',
+    password varchar(20) not null comment '회원 비밀번호',
+    name     varchar(20) not null comment '회원 이름',
+    phone    varchar(20) not null comment '회원 전화번호',
+)
