--- src/main/java/nextstep/auth/AuthenticationException.java
@@ -1,4 +1,0 @@
-package nextstep.auth;
-
-public class AuthenticationException extends RuntimeException {
-}
