--- src/main/java/nextstep/auth/TokenResponse.java
@@ -1,16 +1,0 @@
-package nextstep.auth;
-
-public class TokenResponse {
-    public String accessToken;
-
-    public TokenResponse() {
-    }
-
-    public TokenResponse(String accessToken) {
-        this.accessToken = accessToken;
-    }
-
-    public String getAccessToken() {
-        return accessToken;
-    }
-}
