--- src/main/java/nextstep/schedule/Schedule.java
@@ -1,45 +1,0 @@
-package nextstep.schedule;
-
-import nextstep.theme.Theme;
-
-import java.time.LocalDate;
-import java.time.LocalTime;
-
-public class Schedule {
-    private Long id;
-    private Theme theme;
-    private LocalDate date;
-    private LocalTime time;
-
-    public Schedule() {
-    }
-
-    public Schedule(Long id, Theme theme, LocalDate date, LocalTime time) {
-        this.id = id;
-        this.theme = theme;
-        this.date = date;
-        this.time = time;
-    }
-
-    public Schedule(Theme theme, LocalDate date, LocalTime time) {
-        this.theme = theme;
-        this.date = date;
-        this.time = time;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public Theme getTheme() {
-        return theme;
-    }
-
-    public LocalDate getDate() {
-        return date;
-    }
-
-    public LocalTime getTime() {
-        return time;
-    }
-}
