--- src/main/java/nextstep/schedule/ScheduleService.java
@@ -1,31 +1,0 @@
-package nextstep.schedule;
-
-import nextstep.theme.Theme;
-import nextstep.theme.ThemeDao;
-import org.springframework.stereotype.Service;
-
-import java.util.List;
-
-@Service
-public class ScheduleService {
-    private ScheduleDao scheduleDao;
-    private ThemeDao themeDao;
-
-    public ScheduleService(ScheduleDao scheduleDao, ThemeDao themeDao) {
-        this.scheduleDao = scheduleDao;
-        this.themeDao = themeDao;
-    }
-
-    public Long create(ScheduleRequest scheduleRequest) {
-        Theme theme = themeDao.findById(scheduleRequest.getThemeId());
-        return scheduleDao.save(scheduleRequest.toEntity(theme));
-    }
-
-    public List<Schedule> findByThemeIdAndDate(Long themeId, String date) {
-        return scheduleDao.findByThemeIdAndDate(themeId, date);
-    }
-
-    public void deleteById(Long id) {
-        scheduleDao.deleteById(id);
-    }
-}
