--- src/main/java/nextstep/support/DuplicateEntityException.java
@@ -1,4 +1,0 @@
-package nextstep.support;
-
-public class DuplicateEntityException extends RuntimeException {
-}
