--- src/main/java/nextstep/theme/ThemeController.java
@@ -1,36 +1,0 @@
-package nextstep.theme;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import java.net.URI;
-import java.util.List;
-
-@RestController
-@RequestMapping("/themes")
-public class ThemeController {
-    private ThemeService themeService;
-
-    public ThemeController(ThemeService themeService) {
-        this.themeService = themeService;
-    }
-
-    @PostMapping
-    public ResponseEntity<Void> createTheme(@RequestBody ThemeRequest themeRequest) {
-        Long id = themeService.create(themeRequest);
-        return ResponseEntity.created(URI.create("/themes/" + id)).build();
-    }
-
-    @GetMapping
-    public ResponseEntity<List<ThemeResponse>> showThemes() {
-        List<ThemeResponse> results = themeService.findAll();
-        return ResponseEntity.ok().body(results);
-    }
-
-    @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteTheme(@PathVariable Long id) {
-        themeService.delete(id);
-
-        return ResponseEntity.noContent().build();
-    }
-}
