--- src/main/java/nextstep/theme/ThemeService.java
@@ -1,35 +1,0 @@
-package nextstep.theme;
-
-import org.springframework.stereotype.Service;
-
-import java.util.List;
-import java.util.stream.Collectors;
-
-@Service
-public class ThemeService {
-    private ThemeDao themeDao;
-
-    public ThemeService(ThemeDao themeDao) {
-        this.themeDao = themeDao;
-    }
-
-    public Long create(ThemeRequest themeRequest) {
-        return themeDao.save(themeRequest.toEntity());
-    }
-
-    public List<ThemeResponse> findAll() {
-        List<Theme> themes = themeDao.findAll();
-        return themes.stream()
-                .map(it -> new ThemeResponse(it.getId(), it.getName(), it.getDesc(), it.getPrice()))
-                .collect(Collectors.toList());
-    }
-
-    public void delete(Long id) {
-        Theme theme = themeDao.findById(id);
-        if (theme == null) {
-            throw new NullPointerException();
-        }
-
-        themeDao.delete(id);
-    }
-}
