--- src/main/resources/application.properties
@@ -1,2 +1,0 @@
-security.jwt.token.secret-key= learning-test-spring
-security.jwt.token.expire-length= 3600000
